flaskApp_token/backend/requirements.txt
Flask>=2.0.2,<3.0.0
Flask-RESTful>=0.3.9,<1.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-Migrate>=3.0.0,<4.0.0
SQLAlchemy<2.0.0
psycopg2-binary>=2.9.1,<3.0.0
python-dotenv>=0.19.1,<1.0.0
flask_cors>=3.0.0,<4.0.0
flask_jwt_extended==4.3.0
pymysql
requests


